(GH-1) Make hugo-calendly-shortcode a hugo module
Prior to this commit, the shortcode was only installable via either
git submodule or copy-paste methods; this commit turns the theme
into a hugo module so that it can be installed/updated/removed more
idiomatically.

This commit also adds documentation for installation, update, and
removal of the hugo-calendly-shortcode via hugo modules.